+++ src/onegov.town/onegov/town/models/__init__.py
@@ -2,7 +2,7 @@ from onegov.town.models.atoz import AtoZ
 from onegov.town.models.clipboard import Clipboard
 from onegov.town.models.editor import Editor
 from onegov.town.models.file import File, FileCollection
-from onegov.town.models.form import BuiltinFormDefinition, CustomFormDefinition
+from onegov.town.models.form import CustomFormDefinition
 from onegov.town.models.image import Image, ImageCollection, Thumbnail
 from onegov.town.models.page import AtoZPages, News, Topic
 from onegov.town.models.page_move import PageMove
@@ -22,7 +22,6 @@ from onegov.town.models.traitinfo import TraitInfo
 __all__ = [
     'AtoZ',
     'AtoZPages',
-    'BuiltinFormDefinition',
     'Clipboard',
     'CustomFormDefinition',
     'DaypassResource',